+++ addon/locales/ru/components/fd-color-setting-panel.js
@@ -1,0 +1,6 @@
+export default {
+    'button-caption': 'Применить',
+    'text-color-label': 'Цвет текста',
+    'background-color-label': 'Цвет фона',
+    'stereotype-label': 'Стереотип'
+};